app/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.12.0
numpy>=1.26.0
PyMuPDF==1.25.0
google-genai==1.41.0
openai==1.41.0
